pyproject.toml
[tool.poetry]
name = "analyzer"
version = "0.1.0"
description = "Model-Based Analysis of Forensic-Ready Software Systems"
authors = ["SofijaMaksovic <maksovic@mail.muni.cz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
zope-interface = "^5.5.2"
z3-solver = "^4.11.2.0"
pytest = "^7.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
